HubSpotIntegration/requirements.txt
python-telegram
telebot
config
requests
python-dotenv